Simple Boolean logic operations. Assume we have four input data samples {(0, 0), (0, 1), (1, 0), (1, 1)} to a model. Perform the followings in Python:
Plot the four input data samples. Label the axes and title as appropriate.
Write a function that takes two arguments as an input and performs AND operation. The function returns the result to the main program.
Write two additional functions like in part (b); each would perform OR and XOR operations and return the result to the main program.
Create a for loop to call the three functions each for data sample each time. 
Store the output of each function in a separate list/array.
Print out the results for the three functions at the end of your program.
Note: do not use the internal Boolean (AND, OR, XOR) functions in Python; instead, utilize conditional statements to perform the desired operations.